+++ src/main/kotlin/com/pwlimaverde/return_success_or_error_kt/core/ReturnSuccessOrError.kt
@@ -1,0 +1,3 @@
+package com.pwlimaverde.return_success_or_error_kt.core
+
+sealed interface ReturnSuccessOrError<R>