+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+[1.2.0] - 01/09/2024.
+1 - Correção ortografica README.
+2 - Implementação das Classes ```ReturnSuccessOrError```, ```ErrorReturn```, ```SuccessReturn```,
+```AppError```, ```ErrorGeneric```, ```NoParams```, ```DataSource```, ```UseCaseCallData``` e 
+```ParametersReturnResult```, assim como suas respectivas Classes test.
+
 [1.1.0] - 29/08/2024.
 Inicio do desenvolvimento